updated auth-flow example
