+++ examples/auth-flow/app.js
@@ -1,25 +1,24 @@
 var React = require('react');
-var HashHistory = require('react-router/HashHistory');
-var { createRouter, Route, Link } = require('react-router');
+var { HashHistory, Router, Route, Link, Navigation } = require('react-router');
+var auth = require('./auth');
 
-class App extends React.Component {
-  constructor (props) {
-    super(props);
-    this.state = {
+var App = React.createClass({
+  getInitialState() {
+    return {
       loggedIn: auth.loggedIn()
     };
-  }
+  },
 
   setStateOnAuth (loggedIn) {
     this.setState({
       loggedIn: loggedIn
     });
-  }
+  },
 
   componentWillMount () {
-    auth.onChange = this.setStateOnAuth.bind(this);
+    auth.onChange = this.setStateOnAuth;
     auth.login();
-  }
+  },
 
   render () {
     return (
@@ -39,14 +38,9 @@ class App extends React.Component {
       </div>
     );
   }
-}
+});
 
-function requireAuth (router, state) {
-  if (!auth.loggedIn())
-    router.replaceWith('/login', {}, {'nextPath' : state.location.path});
-}
-
-class Dashboard extends React.Component {
+var Dashboard = React.createClass({
   render () {
     var token = auth.getToken();
     return (
@@ -57,34 +51,33 @@ class Dashboard extends React.Component {
       </div>
     );
   }
-}
+});
 
-class Login extends React.Component {
+var Login = React.createClass({
 
-  constructor (props) {
-    super(props);
-    this.handleSubmit = this.handleSubmit.bind(this);
-    this.state = {
+  mixins: [ Navigation ],
+
+  getInitialState() {
+    return {
       error: false
     };
-  }
+  },
 
   handleSubmit (event) {
     event.preventDefault();
-    var { router } = this.context;
-    var nextPath = router.getCurrentQuery().nextPath;
+    var { nextPath } = this.props.query;
     var email = this.refs.email.getDOMNode().value;
     var pass = this.refs.pass.getDOMNode().value;
     auth.login(email, pass, (loggedIn) => {
       if (!loggedIn)
         return this.setState({ error: true });
       if (nextPath) {
-        router.replaceWith(nextPath);
+        this.replaceWith(nextPath);
       } else {
-        router.replaceWith('/about');
+        this.replaceWith('/about');
       }
     });
-  }
+  },
 
   render () {
     return (
@@ -98,90 +91,36 @@ class Login extends React.Component {
       </form>
     );
   }
-}
+});
 
-Login.contextTypes = {
-  router: React.PropTypes.func
-};
-
-class About extends React.Component {
+var About = React.createClass({
   render () {
     return <h1>About</h1>;
   }
-}
+});
 
-class Logout extends React.Component {
+var Logout = React.createClass({
   componentDidMount () {
     auth.logout();
-  }
+  },
 
   render () {
     return <p>You are now logged out</p>;
   }
+});
+
+function requireAuth (state, router) {
+  if (!auth.loggedIn())
+    router.replaceWith('/login', {nextPath: state.location.path});
 }
 
-
-// Fake authentication lib
-
-var auth = {
-  login (email, pass, cb) {
-    cb = arguments[arguments.length - 1];
-    if (localStorage.token) {
-      if (cb) cb(true);
-      this.onChange(true);
-      return;
-    }
-    pretendRequest(email, pass, (res) => {
-      if (res.authenticated) {
-        localStorage.token = res.token;
-        if (cb) cb(true);
-        this.onChange(true);
-      } else {
-        if (cb) cb(false);
-        this.onChange(false);
-      }
-    });
-  },
-
-  getToken: function () {
-    return localStorage.token;
-  },
-
-  logout: function (cb) {
-    delete localStorage.token;
-    if (cb) cb();
-    this.onChange(false);
-  },
-
-  loggedIn: function () {
-    return !!localStorage.token;
-  },
-
-  onChange: function () {}
-};
-
-function pretendRequest(email, pass, cb) {
-  setTimeout(() => {
-    if (email === 'joe@example.com' && pass === 'password1') {
-      cb({
-        authenticated: true,
-        token: Math.random().toString(36).substring(7)
-      });
-    } else {
-      cb({authenticated: false});
-    }
-  }, 0);
-}
-
-var routes = (
-  <Route component={App}>
-    <Route name="login" component={Login}/>
-    <Route name="logout" component={Logout}/>
-    <Route name="about" component={About}/>
-    <Route name="dashboard" component={Dashboard} onEnter={requireAuth}/>
-  </Route>
-);
-
-var Router = createRouter(routes);
-
-React.render(<Router history={HashHistory}/>, document.getElementById('example'));
+React.render((
+  <Router history={HashHistory}>
+    <Route path="/" component={App}>
+      <Route path="login" component={Login}/>
+      <Route path="logout" component={Logout}/>
+      <Route path="about" component={About}/>
+      <Route path="dashboard" component={Dashboard} onEnter={requireAuth}/>
+    </Route>
+  </Router>
+), document.getElementById('example'));
